Optimize via awk, and add `-h|--help` to help()
Now `awk` can take of help almost entirely on its own, saving the need
for `sed`, but I've also added `-h|--help` to `help()` with description
text shown in `ls --help`, since it looks as though you were going for
a traditional approach here.

I've also made `awk` use `-P` to adhere to POSIX, in-case it's desired.

As long as the formatting is the same as is already present in the case
statement, there shouldn't be any issues with additional file formats.