term: use a custom socket for -f
The -f argument is used to start a terminal that can be waited on (e.g.
when opening an editor in it).  When this alacritty process suddenly
manages several terminal windows it will not shut down when the original
window is closed.  This breaks the "open in editor" workflow in e.g.
alot or qutebrowser.